requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosmtplib==5.0.0
alembic==1.17.2
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.4
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
google-api-core==2.27.0
google-auth==2.41.1
google-geo-type==0.5.0
google-maps-routeoptimization==0.2.0
google-maps-routing==0.8.0
googleapis-common-protos==1.71.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
multidict==6.7.0
packaging==25.0
phonenumbers==9.0.24
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-flask==1.3.0
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.15.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
simple-websocket==1.1.0
SQLAlchemy==2.0.44
twilio==9.8.4
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.3.2
yarl==1.22.0
